requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.2
pdf2image==1.17.0
Pillow==10.4.0
streamlit==1.38.0
typing-extensions==4.12.2
pydantic==2.9.1
pydantic-settings==2.4.0
tenacity==8.3.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
pytest==8.3.2
pytest-asyncio==0.23.8
